Adicionado um novo parametro no body do pix, e feita uma refatoração no componente de confirm order, foi separada a logica do componente
